requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
filelock==3.12.3
fsspec==2023.6.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pydantic==2.3.0
pydantic-core==2.6.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
sniffio==1.3.0
starlette==0.27.0
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
torchvision==0.15.2+cpu
tqdm==4.66.1
transformers==4.32.1
typing-extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
